Rename to standard servo
